Issue #439 Address PR review
