+++ src/components/header_menu/header_menu_component.tsx
@@ -112,17 +112,17 @@ const AboutListItems = (props: Props): JSX.Element => {
             <AboutItem
                 icon={<AboutIcon name='mobile' fontSize={35} marginRight={10}/>}
                 text={<Trans>About Arrival Advisor</Trans>}
-                onPress={buildAboutItemOnPress(props, Routes.About)}
+                onPress={buildOnPressForRoute(props, Routes.About)}
             />
             <AboutItem
                 icon={<AboutIcon name='file' fontSize={20} marginRight={7}/>}
                 text={<Trans>Disclaimer</Trans>}
-                onPress={buildAboutItemOnPress(props, Routes.Disclaimer)}
+                onPress={buildOnPressForRoute(props, Routes.Disclaimer)}
             />
             <AboutItem
                 icon={<AboutIcon name='lock' fontSize={30} marginRight={5}/>}
                 text={<Trans>Privacy policy</Trans>}
-                onPress={(): void => openURL('https://peacegeeks.org/privacy')}
+                onPress={buildOnPressForURL('https://peacegeeks.org/privacy')}
             />
         </View>
     );
@@ -167,11 +167,15 @@ const AboutItem = (props: { readonly icon: JSX.Element, readonly text: JSX.Eleme
         </TouchableOpacity>
     );
 
-const buildAboutItemOnPress = (props: Props, route: Routes): () => void => (
+const buildOnPressForRoute = (props: Props, route: Routes): () => void => (
     (): void => {
         props.closeMenu(); // tslint:disable-line:no-expression-statement
         goToRouteWithoutParameter(route, props.history)(); // tslint:disable-line:no-expression-statement
     }
+);
+
+const buildOnPressForURL = (url: string): () => void => (
+    (): void => openURL(url)
 );
 
 const styles = StyleSheet.create({
